Refactor Convex functions into feature-based modules
Moved and reorganized Convex function files into feature-based directories (jobs, mentors, messages, users). Queries and mutations are now separated for each feature, and index files re-export them for easier imports. Removed legacy application and jobApplications functions, and added helpers for user profile completion.
