+++ convex/jobs/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Jobs Feature - Convex API
+ * Re-exports all jobs queries and mutations
+ */
+
+export * from "./queries";
+export * from "./mutations";
