+++ convex/mentors/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Mentors Feature - Convex API
+ * Re-exports all mentors queries and mutations
+ */
+
+export * from "./queries";
+export * from "./mutations";
