+++ convex/messages/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Messages Feature - Convex API
+ * Re-exports all messages queries and mutations
+ */
+
+export * from "./queries";
+export * from "./mutations";
