+++ convex/users/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Users Feature - Convex API
+ * Re-exports all users queries and mutations
+ */
+
+export * from "./queries";
+export * from "./mutations";
